+++ spec/support/shared_examples/behaves_like_valid_activity_xml.rb
@@ -6,29 +6,30 @@ RSpec.shared_examples "valid activity XML" do
 
   it "contains the activity XML" do
     visit organisation_activity_path(organisation, activity, format: :xml)
-    expect(xml.at("iati-activity/@default-currency").text).to eq(activity.default_currency)
-    expect(xml.at("iati-activity/iati-identifier").text).to eq(activity.identifier)
+
+    expect(xml.at("iati-activity/@default-currency").text).to eq(activity_presenter.default_currency)
+    expect(xml.at("iati-activity/iati-identifier").text).to eq(activity_presenter.iati_identifier)
   end
 
   it "contains the funding organisation XML" do
     visit organisation_activity_path(organisation, activity, format: :xml)
-    expect(xml.at("iati-activity/participating-org[@role = '1']/@ref").text).to eq(activity.funding_organisation_reference)
-    expect(xml.at("iati-activity/participating-org[@role = '1']/@type").text).to eq(activity.funding_organisation_type)
-    expect(xml.at("iati-activity/participating-org[@role = '1']/narrative").text).to eq(activity.funding_organisation_name)
+    expect(xml.at("iati-activity/participating-org[@role = '1']/@ref").text).to eq(activity_presenter.funding_organisation_reference)
+    expect(xml.at("iati-activity/participating-org[@role = '1']/@type").text).to eq(activity_presenter.funding_organisation_type)
+    expect(xml.at("iati-activity/participating-org[@role = '1']/narrative").text).to eq(activity_presenter.funding_organisation_name)
   end
 
   it "contains the accountable organisation XML" do
     visit organisation_activity_path(organisation, activity, format: :xml)
-    expect(xml.at("iati-activity/participating-org[@role = '2']/@ref").text).to eq(activity.accountable_organisation_reference)
-    expect(xml.at("iati-activity/participating-org[@role = '2']/@type").text).to eq(activity.accountable_organisation_type)
-    expect(xml.at("iati-activity/participating-org[@role = '2']/narrative").text).to eq(activity.accountable_organisation_name)
+    expect(xml.at("iati-activity/participating-org[@role = '2']/@ref").text).to eq(activity_presenter.accountable_organisation_reference)
+    expect(xml.at("iati-activity/participating-org[@role = '2']/@type").text).to eq(activity_presenter.accountable_organisation_type)
+    expect(xml.at("iati-activity/participating-org[@role = '2']/narrative").text).to eq(activity_presenter.accountable_organisation_name)
   end
 
   it "contains the extending organisation XML" do
     visit organisation_activity_path(organisation, activity, format: :xml)
-    expect(xml.at("iati-activity/participating-org[@role = '3']/@ref").text).to eq(activity.extending_organisation.iati_reference)
-    expect(xml.at("iati-activity/participating-org[@role = '3']/@type").text).to eq(activity.extending_organisation.organisation_type)
-    expect(xml.at("iati-activity/participating-org[@role = '3']/narrative").text).to eq(activity.extending_organisation.name)
+    expect(xml.at("iati-activity/participating-org[@role = '3']/@ref").text).to eq(activity_presenter.extending_organisation.iati_reference)
+    expect(xml.at("iati-activity/participating-org[@role = '3']/@type").text).to eq(activity_presenter.extending_organisation.organisation_type)
+    expect(xml.at("iati-activity/participating-org[@role = '3']/narrative").text).to eq(activity_presenter.extending_organisation.name)
   end
 
   it "contains the implementing organisations XML" do