+++ CHANGELOG.md
@@ -55,3 +55,5 @@
 - Users can view implementing organisations in the Activity XML
 - Planned start and end dates are mandatory
 - Actual start and end dates must not be in the future
+- Activity XML includes 'iati-identifier' which is includes the reporting organisation
+- Activity XML includes 'iati-identifier' with an identifier that consists of all present levels of hierarchy: fund and/or programme